pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "globaleaks_eph_fs"
version = "0.0.1"
authors = [
  { name="Giovanni Pellerano", email="giovanni.pellerano@globaleaks.org" },
]
description = "An ephemeral ChaCha20-encrypted filesystem implementation using fusepy and cryptography suitable for privacy-sensitive applications, such as whistleblowing platforms. "
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]

dependencies = [
    "cryptography",
    "fusepy"
]

[project.urls]
Homepage = "https://github.com/globaleaks/globaleaks-eph-fs"
Issues = "https://github.com/globaleaks/globaleaks-eph-fs/issues"


[project.entry-points."console_scripts"]
globaleaks-eph-fs = "globaleaks_eph_fs:main"
